requirements.txt
accelerate==1.12.0
aiokafka==0.13.0
annotated-doc==0.0.4
anyio==4.12.1
async-timeout==5.0.1
audioread==3.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
filelock==3.25.0
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kafka-python==2.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
packaging==26.0
peft==0.9.0
platformdirs==4.9.2
pooch==1.9.0
psutil==7.2.2
pycparser==3.0
Pygments==2.19.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.10.0
torchaudio==2.10.0
tqdm==4.67.3
transformers==4.40.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.7
